Introduce new design
CSS Grid are very usefull to create a layout with a left sidebar and content in the rest of the space.
Footer at bottom use every width space.

This design keep a simple sidebar with import hyperlinks and act as a single infinite vertical scroll page,
with the power of HTML and CSS!

Signed-off-by: Max Charrier <max@puffer.fish>